Moved target selection and damage application to targetcomponent script to better keep track of which target is selected

+++ Assets/Scripts/WeaponScriptableObject.cs
@@ -16,6 +16,6 @@ public class WeaponScriptableObject : ScriptableObject
 {
     public DamageType damageType;
     public string weaponName;
-    public float baseDamage;
+    public int baseDamage;
     public Sprite weaponSprite;
 }
